データ入力画面

+++ lib/charts/data_class.dart
@@ -1,14 +1,14 @@
 class DataPoint {
-  late double value;
+  late int value;
   late DateTime time;
   DataPoint(this.value, this.time);
 
   DataPoint.fromMap(Map map) {
     time = map['timestamp'].toDate();
-    if (map['value'] is double) {
+    if (map['value'] is int) {
       value = map['value'];
     } else {
-      value = (map['value'] as int).toDouble();
+      value = (map['value'] as double).toInt();
     }
   }
